+++ src/app/pages/pages-menu.ts
@@ -1,7 +1,7 @@
 import { NbMenuItem } from '@nebular/theme';
 import { CommonData } from '../common/common';
 
-export const MENU_ITEMS: NbMenuItem[] = [
+export const MENU_ITEMS_ALL: NbMenuItem[] = [
   {
     title: 'Bán thuốc',
     icon: 'shopping-cart-outline',
@@ -34,6 +34,30 @@ export const MENU_ITEMS: NbMenuItem[] = [
 
     ]
   },
-  
- 
+];
+export const MENU_ITEMS_STAFF: NbMenuItem[] = [
+  {
+    title: 'Bán thuốc',
+    icon: 'shopping-cart-outline',
+    link: '/'+CommonData.StorePagePrefix+'/sell-medicine',
+    home: true,
+  },
+  {
+    title: 'Nhập thuốc',
+    icon: 'shopping-cart-outline',
+    link: '/'+CommonData.StorePagePrefix+'/import-medicine',
+    home: true,
+  },
+  {
+    title: 'Báo cáo',
+    icon: 'shopping-cart-outline',
+      children: [
+     
+      {
+        title: 'Lịch sử bán hàng',
+        link:  '/'+CommonData.StorePagePrefix+'/report/myhistory-sell',
+      },
+
+    ]
+  },
 ];
