+++ src/app/pages/pages-routing.module.ts
@@ -19,20 +19,11 @@ const routes: Routes = [{
       component: CreateNewComponent
     },
     {
-      path: 'dashboard',
-      component: ECommerceComponent,
-    },
-
-    {
-      path: 'iot-dashboard',
-      component: DashboardComponent,
-    },
-    {
       path: 'sell-medicine',
       canActivate:[AuthGuardService],
-      // loadChildren: () => import('./sell-medicine/sell-medicine.module')
-      //   .then(m => m.SellMedicineModule),
-      loadChildren:'./sell-medicine/sell-medicine.module#SellMedicineModule'
+      loadChildren: () => import('./sell-medicine/sell-medicine.module')
+        .then(m => m.SellMedicineModule),
+      // loadChildren:'./sell-medicine/sell-medicine.module#SellMedicineModule'
     },
     {
       path: 'import-medicine',
@@ -45,56 +36,6 @@ const routes: Routes = [{
       canActivate:[AuthGuardService],
       loadChildren: () => import('./report/report.module')
         .then(m => m.ReportModule),
-    },
-    {
-      path: 'layout',
-      loadChildren: () => import('./layout/layout.module')
-        .then(m => m.LayoutModule),
-    },
-    {
-      path: 'forms',
-      loadChildren: () => import('./forms/forms.module')
-        .then(m => m.FormsModule),
-    },
-    {
-      path: 'ui-features',
-      loadChildren: () => import('./ui-features/ui-features.module')
-        .then(m => m.UiFeaturesModule),
-    },
-    {
-      path: 'modal-overlays',
-      loadChildren: () => import('./modal-overlays/modal-overlays.module')
-        .then(m => m.ModalOverlaysModule),
-    },
-    {
-      path: 'extra-components',
-      loadChildren: () => import('./extra-components/extra-components.module')
-        .then(m => m.ExtraComponentsModule),
-    },
-    {
-      path: 'maps',
-      loadChildren: () => import('./maps/maps.module')
-        .then(m => m.MapsModule),
-    },
-    {
-      path: 'charts',
-      loadChildren: () => import('./charts/charts.module')
-        .then(m => m.ChartsModule),
-    },
-    {
-      path: 'editors',
-      loadChildren: () => import('./editors/editors.module')
-        .then(m => m.EditorsModule),
-    },
-    {
-      path: 'tables',
-      loadChildren: () => import('./tables/tables.module')
-        .then(m => m.TablesModule),
-    },
-    {
-      path: 'miscellaneous',
-      loadChildren: () => import('./miscellaneous/miscellaneous.module')
-        .then(m => m.MiscellaneousModule),
     },
     {
       path: '',
