+++ src/app/pages/pages.component.ts
@@ -1,6 +1,7 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 
-import { MENU_ITEMS } from './pages-menu';
+import { MENU_ITEMS_ALL,MENU_ITEMS_STAFF } from './pages-menu';
+import { TokenStorageService } from '../service/tokenstorage.service';
 
 @Component({
   selector: 'ngx-pages',
@@ -12,7 +13,24 @@ import { MENU_ITEMS } from './pages-menu';
     </ngx-one-column-layout>
   `,
 })
-export class PagesComponent {
-  menu = MENU_ITEMS;
+export class PagesComponent implements OnInit {
+
+  menu ;
+  constructor(
+    private tokenStorageServce: TokenStorageService,
+  ) {
+    console.log("Check role");
+    let role = this.tokenStorageServce.getRole();
+    if(role==='ROLE_ADMIN'){
+      
+      this.menu=MENU_ITEMS_ALL;
+    }else{
+      this.menu=MENU_ITEMS_STAFF;
+    }
+  }
+  ngOnInit(): void {
+   
+
+  }
 }
