+++ src/app/pages/report/my-sell-history/my-sell-history.component.html
@@ -1,0 +1,56 @@
+<kendo-grid
+[data]="view | async"
+(cellClick)="cellClickEvent($event)"
+[kendoGridBinding]="products" [pageable]="true" [pageSize]="100">
+  <ng-template kendoGridToolbarTemplate>
+    <div class="row">
+      <div class="col-md-4">
+        <label for="fromDate">Từ ngày:</label>
+        <input id="fromDate" [(ngModel)]="fromDate" nbInput placeholder="Ngày nhập thuốc" [nbDatepicker]="formpicker">
+        <nb-datepicker #formpicker format="dd\MM\yyyy"></nb-datepicker>
+      </div>
+      <div class="col-md-5">
+        <label for="toDate">Đến ngày:</label>
+        <input id="toDate" [(ngModel)]="toDate" nbInput placeholder="Ngày nhập thuốc" [nbDatepicker]="formpickertoDate">
+        <nb-datepicker #formpickertoDate format="dd\MM\yyyy"></nb-datepicker>
+      </div>
+      <div class="col-md-3">
+        <button nbButton status="danger" (click)="testPrintDate()"> Truy xuất </button>
+        <button kendoGridPDFCommand icon="file-pdf">In</button>
+      </div>
+    </div>
+  </ng-template>
+  <kendo-grid-column field="id" title="MHĐ" width="70">
+  </kendo-grid-column>
+  <kendo-grid-column field="sellDate" title="Ngày bán">
+  </kendo-grid-column>
+  <kendo-grid-column field="staffName" title="Người bán">
+  </kendo-grid-column>
+  <kendo-grid-column field="customerName" title="Tên khách mua">
+  </kendo-grid-column>
+  <kendo-grid-column field="total"   [format]="formatOptions" title="Tổng tiền">
+  </kendo-grid-column>
+
+  <!-- <kendo-grid-column field="UnitsInStock" title="In stock" width="80">
+  </kendo-grid-column>
+  <kendo-grid-column field="Discontinued" title="Discontinued" width="120">
+      <ng-template kendoGridCellTemplate let-dataItem>
+          <input type="checkbox" [checked]="dataItem.Discontinued"/>
+      </ng-template>
+  </kendo-grid-column> -->
+  <kendo-grid-pdf fileName="{{genFileReport()}}" [allPages]="true" paperSize="A4" [repeatHeaders]="true"
+    [landscape]="true">
+    <kendo-grid-pdf-margin top="2cm" right="1cm" left="1cm" bottom="2cm"></kendo-grid-pdf-margin>
+    <ng-template kendoGridPDFTemplate let-pageNum="pageNum" let-totalPages="totalPages">
+      <div class="page-template">
+        <div class="header">
+          <div style="float: right">Trang {{ pageNum }} / {{ totalPages }}</div>
+          Nhà thuốc thú Y Lan Kha
+        </div>
+        <div class="footer">
+          Trang {{ pageNum }} / {{ totalPages }}
+        </div>
+      </div>
+    </ng-template>
+  </kendo-grid-pdf>
+</kendo-grid>
