+++ src/app/pages/sell-medicine/sell/sell.component.html
@@ -1,0 +1,99 @@
+<div class="row">
+  <div class="col-md-8">
+    <nb-card>
+      <nb-card-header>
+        Hóa đơn bán thuốc
+        <!-- <ngx-button></ngx-button> -->
+      </nb-card-header>
+      <nb-card-body>
+        <!-- <form> -->
+        <nb-card>
+          <nb-card-header>Thông tin khách hàng</nb-card-header>
+          <nb-card-body>
+            <div class="row">
+              <div class="col-6 col-md-4">
+                <!-- <span>Hình thức mua hàng: </span>
+                <nb-select selected="other" (selectedChange)="selectType($event)">
+                  <nb-option value="company">Công ty</nb-option>
+                  <nb-option value="farm">Trang trại</nb-option>
+                  <nb-option value="other">Khách lẻ</nb-option>
+                </nb-select> -->
+                <span style="font-size: 10px;margin-right: 5px;">Công ty Cổ phần Emoss</span>
+                <nb-checkbox [(checked)]="isEmoss" (change)="changeToEmossCompany()" status="success"></nb-checkbox>
+              </div>
+              <div class="col-6 col-md-4">
+                <input [(ngModel)]="customer.phoneNumber" (change)="changePhone()" list="listCustomerPhone" type="text" nbInput
+                  placeholder="Số điện thoại">
+                <datalist id="listCustomerPhone" >
+                  <option *ngFor="let phone of listPhone" value="{{phone}}">
+                </datalist>
+              </div>
+              <div class="col-6 col-md-4">
+                <input [(ngModel)]="customer.name" (change)="changeName()" list="listCustomerName" type="text" nbInput placeholder="Tên khách">
+                <datalist id="listCustomerName" >
+                  <option *ngFor="let name of listName" value="{{name}}">
+                </datalist>
+              </div>
+            </div>
+            <div class="row" style="margin-top: 10px;">
+              <div class="col-6 col-md-4">
+              </div>
+              <div class="col-6 col-md-4">
+              </div>
+              <div class="col-6 col-md-4">
+                <input [(ngModel)]="customer.traiDungThuoc" type="text" nbInput placeholder="Tên chủ trại">
+              </div>
+            </div>
+          </nb-card-body>
+        </nb-card>
+
+      </nb-card-body>
+      <nb-card-body>
+        <nb-card>
+          <nb-card-header>Các loại thuốc</nb-card-header>
+          <nb-card-body>
+            <ng2-smart-table [settings]="settings" [source]="source" (deleteConfirm)="onDeleteConfirm($event)"
+              (edit)="onEditConfirm($event)" (editConfirm)="onEditConfirm($event)"
+              (createConfirm)="onCreateConfirm($event)">
+            </ng2-smart-table>
+          </nb-card-body>
+        </nb-card>
+
+      </nb-card-body>
+      <nb-card-footer>
+        <div class="col-md-1">
+          <button nbButton status="danger" (click)="seeInvoice()"> Test Pop up invoice </button>
+        </div>
+        <div class="row">
+          <div class="col-md-5"></div>
+          <div class="col-md-1">
+            <button nbButton status="danger" (click)="printPdf(null)"> Tạo </button>
+          </div>
+          <div class="col-md-1">
+            <button nbButton status="danger" (click)="printPdf(pdf)"> Tạo và In </button>
+          </div>
+          <div class="col-md-1"></div>
+
+          <div class="col-md-1">
+
+            <nb-select style="background-color: #252547;" selected="A5" (selectedChange)="changeTypePrint($event)">
+              <nb-option value="A3">Khổ A3</nb-option>
+              <nb-option value="A4">Khổ A4</nb-option>
+              <nb-option value="A5">Khổ A5</nb-option>
+            </nb-select>
+          </div>
+          <div class="col-md-1" style="margin-left: 20px;">
+            <nb-checkbox [(checked)]="autoCheckDate" status="success"></nb-checkbox>
+          </div>
+        </div>
+
+      </nb-card-footer>
+    </nb-card>
+  </div>
+  <div class="col-md-4">
+    <kendo-pdf-export #pdf paperSize="{{paperSize}}" margin="1cm">
+      <my-invoice [data]="dataInvoiceRow" [autoCheckDate]="autoCheckDate" [isEmoss]="isEmoss" [(customer)]="customer">
+      </my-invoice>
+    </kendo-pdf-export>
+  </div>
+</div>
