+++ src/typescript-angular-client/api/medicineController.service.ts
@@ -240,6 +240,6 @@ export class MedicineControllerService {
             }
         );
     }
-   
+
 
 }
