+++ src/typescript-angular-client/model/invoiceDto.ts
@@ -1,0 +1,7 @@
+export interface InvoiceDto{
+    productName?:string;
+    unit?:string;
+    unitPrice?:number;
+    amount?:number;
+    addMore?:number;
+}
